+++ HotelBookingSystem/HotelBookingSystem.Domain/Dto/HotelClientPostDto.cs
@@ -1,15 +1,10 @@
-﻿namespace HotelBookingSystem.Api.Dto;
+﻿namespace HotelBookingSystem.Domain.Dto;
 
 /// <summary>
-/// DTO class for GET method for hotel client
+/// DTO class for POST method for hotel client
 /// </summary>
-public class HotelClientGetDto
+public class HotelClientPostDto
 {
-    /// <summary>
-    /// Client ID
-    /// </summary>
-    public required int Id { get; set; }
-
     /// <summary>
     /// The number of passport of the lodger
     /// </summary>
@@ -30,8 +25,13 @@ public class HotelClientGetDto
     /// </summary>
     public string? Patronymic { get; set; }
 
-    /// <summary>
+    /*/// <summary>
     /// Client's date of birth
     /// </summary>
-    public required DateOnly Birthdate { get; set; }
+    public required DateOnly Birthdate { get; set; }*/
+
+    /// <summary>
+    /// Client's date of birth in "yyyy-MM-dd" format
+    /// </summary>
+    public required string Birthdate { get; set; }
 }
